Split C++'s smoke_test into 4 stand alone ones.
